Updated Project Properties for Libraries
Made the project library for `swingx-1.6.1.jar` relative (thanks @jkovalsky for showing me how!) and made the
`fileNewEmployeeMenuItem` not visible, since that window does nothing. 

The code that I wanted to test is behind the `DepartmentBrowser` `JInternalFrame` window. Please look at this functionality and give me your honest thoughts on how it works. **However**, please pay no attention to the code formatting for this test, as I kind of just threw it together in a few hours. If I implement this functionality in LoadMaster™, I will format it much better.

Thank you in advance for your input on this matter.

-SC
Signed-off-by:Sean Carrick <sean at pekinsoft dot com>